Factor out raptorcast UDP protocol
No functional changes
