+++ monad-raptorcast/benches/raptor_bench.rs
@@ -8,7 +8,9 @@ use monad_crypto::hasher::{Hasher, HasherType};
 use monad_dataplane::network::MONAD_GSO_SIZE;
 use monad_raptor::ManagedDecoder;
 use monad_raptorcast::{
-    build_messages, parse_message, BuildTarget, EpochValidators, Validator, SIGNATURE_CACHE_SIZE,
+    udp::{build_messages, parse_message},
+    util::{BuildTarget, EpochValidators, Validator},
+    SIGNATURE_CACHE_SIZE,
 };
 use monad_secp::{KeyPair, SecpSignature};
 use monad_types::{NodeId, Stake};
